requirements.txt
Flask==2.2.5
tensorflow-cpu==2.13.0
Pillow==9.5.0
numpy==1.24.3
h5py==3.10.0
